Split code and definitions into multiple files
