+++ flash/settings.c
@@ -1,0 +1,5 @@
+#include "settings.h"
+
+u8 palette_selected = 1;
+u8 vel_sensitive = 0;
+u8 top_lights_config = 0; // 0 = PRO, 1 = MK2, 2 = MK2 Rotated, 3 = MK2 Mirrored
