+++ other/modulo.c
@@ -1,0 +1,3 @@
+#include "modulo.h"
+
+#define modulo(x, y) ((x % y + y) % y) // Fix negative modulo behavior
